added tests for backend, and fixes several bugs in it

+++ backend/encryptedbk/replicatedbk.go
@@ -53,12 +53,11 @@ func NewReplicatedBackend(backend backend.Backend, keysFile string, readonly boo
 func (b *ReplicatedBackend) initFromExistingBk() error {
 	log.Infof("Starting with an existing backend. Comparing local and remote keys.")
 
-	localKeys, err := b.GetAllEncryptingKeys()
+	localKeys, err := b.getAllEncryptingKeys()
 	if err != nil {
 		log.Errorf(err.Error())
 		return err
 	}
-
 	if len(localKeys) == 0 {
 		log.Warningf("No local backend encrypting keys were found. Backend will not work until you add encrypting keys")
 		b.readonly = true
@@ -66,6 +65,10 @@ func (b *ReplicatedBackend) initFromExistingBk() error {
 	}
 
 	remoteKeys, err := b.baseBk.GetKeys([]string{rootDir})
+	for i, _ := range remoteKeys {
+		remoteKeys[i] = pathToID(remoteKeys[i])
+	}
+
 	if err != nil {
 		log.Errorf(err.Error())
 		return err
@@ -104,7 +107,7 @@ func (b *ReplicatedBackend) initFromExistingBk() error {
 func (b *ReplicatedBackend) initFromEmptyBk() error {
 	log.Infof("Starting with empty backend")
 
-	localKeys, err := b.GetAllEncryptingKeys()
+	localKeys, err := b.getAllEncryptingKeys()
 	if err != nil {
 		log.Errorf(err.Error())
 		return err
@@ -112,7 +115,7 @@ func (b *ReplicatedBackend) initFromEmptyBk() error {
 
 	if len(localKeys) == 0 {
 		log.Infof("No local backend encrypting keys were found, generating new key 'key0'")
-		return b.GenerateEncryptingKey("key0", false)
+		return b.generateEncryptingKey("key0", false)
 	} else {
 
 		for _, key := range localKeys {
@@ -142,12 +145,15 @@ func (b *ReplicatedBackend) GetKeys(path []string) ([]string, error) {
 	}
 
 	var err error
+	err = trace.Errorf("Backen can't be accessed because there are no valid local encrypting keys")
 
 	for _, bk := range b.ebk {
-		var keys []string
-		keys, err = bk.GetKeys(path)
-		if err == nil {
-			return keys, nil
+		if bk.IsExisting() {
+			var keys []string
+			keys, err = bk.GetKeys(path)
+			if err == nil {
+				return keys, nil
+			}
 		}
 		log.Warningf("Key %s is not valid", bk.KeyID)
 	}
@@ -275,6 +281,8 @@ func (b *ReplicatedBackend) GetValAndTTL(path []string, key string) ([]byte, tim
 }
 
 func (b *ReplicatedBackend) AcquireLock(token string, ttl time.Duration) error {
+	b.mutex.Lock()
+	defer b.mutex.Unlock()
 	return b.baseBk.AcquireLock(token, ttl)
 }
 
@@ -283,6 +291,14 @@ func (b *ReplicatedBackend) ReleaseLock(token string) error {
 }
 
 func (b *ReplicatedBackend) GenerateEncryptingKey(id string, copyData bool) error {
+	b.mutex.Lock()
+	b.baseBk.AcquireLock(lock1, 0)
+	defer b.baseBk.ReleaseLock(lock1)
+	defer b.mutex.Unlock()
+	return b.generateEncryptingKey(id, copyData)
+}
+
+func (b *ReplicatedBackend) generateEncryptingKey(id string, copyData bool) error {
 	if b.readonly {
 		log.Errorf("Can't generate new backend encrypting key without all the encrypting keys used in remote backend. Backend work in readonly mode")
 		return trace.Errorf("Can't generate new backend encrypting key without all the encrypting keys used in remote backend. Backend work in readonly mode")
@@ -296,10 +312,18 @@ func (b *ReplicatedBackend) GenerateEncryptingKey(id string, copyData bool) erro
 	key := Key{}
 	key.ID = id
 	key.Value = keyValue[:]
-	return b.AddEncryptingKey(key, copyData)
+	return b.addEncryptingKey(key, copyData)
 }
 
 func (b *ReplicatedBackend) GetEncryptingKey(id string) (Key, error) {
+	b.mutex.Lock()
+	b.baseBk.AcquireLock(lock1, 0)
+	defer b.baseBk.ReleaseLock(lock1)
+	defer b.mutex.Unlock()
+	return b.getEncryptingKey(id)
+}
+
+func (b *ReplicatedBackend) getEncryptingKey(id string) (Key, error) {
 	value, err := b.keyStorage.GetVal([]string{"values"}, id)
 	if err != nil {
 		if !teleport.IsNotFound(err) {
@@ -314,6 +338,14 @@ func (b *ReplicatedBackend) GetEncryptingKey(id string) (Key, error) {
 }
 
 func (b *ReplicatedBackend) GetAllEncryptingKeys() ([]Key, error) {
+	b.mutex.Lock()
+	b.baseBk.AcquireLock(lock1, 0)
+	defer b.baseBk.ReleaseLock(lock1)
+	defer b.mutex.Unlock()
+	return b.getAllEncryptingKeys()
+}
+
+func (b *ReplicatedBackend) getAllEncryptingKeys() ([]Key, error) {
 	ids, err := b.keyStorage.GetKeys([]string{"values"})
 	if err != nil {
 		log.Errorf(err.Error())
@@ -322,7 +354,7 @@ func (b *ReplicatedBackend) GetAllEncryptingKeys() ([]Key, error) {
 
 	keys := make([]Key, len(ids))
 	for i, id := range ids {
-		keys[i], err = b.GetEncryptingKey(id)
+		keys[i], err = b.getEncryptingKey(id)
 		if err != nil {
 			log.Errorf(err.Error())
 			return nil, err
@@ -333,11 +365,16 @@ func (b *ReplicatedBackend) GetAllEncryptingKeys() ([]Key, error) {
 
 func (b *ReplicatedBackend) AddEncryptingKey(key Key, copyData bool) error {
 	b.mutex.Lock()
+	b.baseBk.AcquireLock(lock1, 0)
+	defer b.baseBk.ReleaseLock(lock1)
 	defer b.mutex.Unlock()
+	return b.addEncryptingKey(key, copyData)
+}
 
-	_, err := b.GetEncryptingKey(key.ID)
+func (b *ReplicatedBackend) addEncryptingKey(key Key, copyData bool) error {
+	_, err := b.getEncryptingKey(key.ID)
 	if err == nil {
-		log.Errorf("Error: Key %s already exists", key.ID)
+		log.Warningf("Error: Key %s already exists", key.ID)
 		return trace.Errorf("Error: Key %s already exists", key.ID)
 	}
 	if !teleport.IsNotFound(err) {
@@ -352,10 +389,12 @@ func (b *ReplicatedBackend) AddEncryptingKey(key Key, copyData bool) error {
 		log.Errorf(err.Error())
 		return err
 	}
-	if copyData {
+	if copyData && len(b.ebk) > 0 && !bk.IsExisting() {
 		err = b.copy(b.ebk[0], bk, []string{})
 		if err != nil {
 			log.Errorf(err.Error())
+			bk.DeleteAll()
+			b.keyStorage.DeleteKey([]string{"values"}, key.ID)
 			return err
 		}
 	}
@@ -366,10 +405,31 @@ func (b *ReplicatedBackend) AddEncryptingKey(key Key, copyData bool) error {
 }
 
 func (b *ReplicatedBackend) DeleteEncryptingKey(id string) error {
+	b.mutex.Lock()
+	b.baseBk.AcquireLock(lock1, 0)
+	defer b.baseBk.ReleaseLock(lock1)
+	defer b.mutex.Unlock()
+
+	anotherValidKey := false
+	for _, bk := range b.ebk {
+		if bk.KeyID != id && bk.IsExisting() {
+			anotherValidKey = true
+		}
+	}
+
+	if !anotherValidKey {
+		log.Warningf("Key %s is the last valid key on this server, it can't be deleted", id)
+		return trace.Errorf("Key %s is the last valid key on this server, it can't be deleted", id)
+	}
+
 	err := b.keyStorage.DeleteKey([]string{"values"}, id)
-	if err != nil {
+	if err != nil && !teleport.IsNotFound(err) {
 		log.Errorf(err.Error())
 		return err
+	}
+
+	if !teleport.IsNotFound(err) {
+		log.Infof("Key %s was deleted from local keys", id)
 	}
 
 	for i, bk := range b.ebk {
@@ -380,15 +440,30 @@ func (b *ReplicatedBackend) DeleteEncryptingKey(id string) error {
 				return err
 			}
 			b.ebk = append(b.ebk[:i], b.ebk[i+1:]...)
+			log.Infof("Key %s was deleted from remote backend keys", id)
 			return nil
 		}
 	}
-	log.Warningf("Deleting encrypting key: key %s doesn't exists in backend", id)
+
+	err = b.baseBk.DeleteBucket([]string{rootDir}, idToPath(id))
+	if err == nil {
+		log.Infof("Key %s was deleted from remote backend keys", id)
+	}
+
 	return nil
 }
 
-func (b *ReplicatedBackend) ListRemoteEncryptingKeys() ([]string, error) {
+func (b *ReplicatedBackend) GetRemoteEncryptingKeys() ([]string, error) {
+	b.mutex.Lock()
+	b.baseBk.AcquireLock(lock1, 0)
+	defer b.baseBk.ReleaseLock(lock1)
+	defer b.mutex.Unlock()
+
 	ids, err := b.baseBk.GetKeys([]string{rootDir})
+	for i, _ := range ids {
+		ids[i] = pathToID(ids[i])
+	}
+
 	if err != nil {
 		log.Errorf(err.Error())
 		return nil, err
